configure.inc: strip trailing characters from 'uname -s' output
On Cygwin, 'uname -s' includes the version of the underlying operating
system; here it is 'CYGWIN_NT-6.0'. The configure script attempts to
define this in config/config.h, but '#define OS_CYGWIN_NT-6.0 1' is an
invalid macro definition.

This truncates the value to just 'CYGWIN', to match $OS in the Makefile.